+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-#CHANGELOG
+# CHANGELOG
 
 This changelog references the relevant changes done between versions.
 
@@ -6,6 +6,7 @@ To get the diff for a specific change, go to https://github.com/LIN3S/AdminBundl
 To get the diff between two versions, go to https://github.com/LIN3S/AdminBundle/compare/v0.2.0...v0.3.0
 
 * 0.5.0
+    * Added `id` getter method in the Entity configuration class.
     * Added novalidate by default in the admin form theme.
     * Added HandleCommandActionType.
     * [BC break] Refactored the whole js and scss files.